Task7 Upd
