+++ src/components/Projects/ProjectList.css
@@ -1,3 +1,22 @@
-.project-list{
-    padding-inline-start: 0
+.project-list {
+  padding-inline-start: 0;
+}
+.projects {
+  padding: 1rem;
+  background-color: rgb(31, 31, 31);
+  margin: 2rem auto;
+  width: 50rem;
+  max-width: 95%;
+  border-radius: 12px;
+  box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
+}
+.projects__caption {
+  display: flex;
+  justify-content: space-between;
+  padding: 0.5rem;
+  margin: 1rem 0;
+  background-color: #4b4b4b;
+  color: white;
+  flex-direction: column;
+  text-align: center;
 }
