--- src/components/Projects/Projects.css
@@ -1,19 +1,0 @@
-.projects {
-  padding: 1rem;
-  background-color: rgb(31, 31, 31);
-  margin: 2rem auto;
-  width: 50rem;
-  max-width: 95%;
-  border-radius: 12px;
-  box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
-}
-.projects__caption {
-  display: flex;
-  justify-content: space-between;
-  padding: 0.5rem;
-  margin: 1rem 0;
-  background-color: #4b4b4b;
-  color:white;
-  flex-direction: column;
-  text-align: center;
-}
