+++ src/components/UI/Button.js
@@ -3,7 +3,7 @@ import "./Button.css";
 
 const Button = (props) => {
   return (
-    <button className="button" onClick={props.onClick}>
+    <button className="button" type={props.type} onClick={props.onClick}>
       {props.children}
     </button>
   );
